Tensorflow: Attempt to support it in AOSP
This is a work in progress.

This PR contains TensorFlow Lite libraries converted for usage within AOSP.

It also contains a tflite model trained to perform nsfw image classification.

This idea is to have tflite model as part of AOSP, use the tensorflow libraries and custom classifier
within framework and intercept any image created by the camera. After classification, we would either return
normal image (if it was neutral) or discard the image if it was NSFW.

This PR goes together with a PR within frameworks/base